refactor: 소스 코드 정리 및 오류 수정
