+++ src/main/java/com/hhplus/tdd/concert/scheduler/ConcertScheduler.java
@@ -1,8 +1,8 @@
 package com.hhplus.tdd.concert.scheduler;
 
+import com.hhplus.tdd.concert.domain.event.MessageProducer;
 import com.hhplus.tdd.concert.domain.model.ConcertEvent;
 import com.hhplus.tdd.concert.domain.model.Outbox;
-import com.hhplus.tdd.concert.domain.kafka.MessageProducer;
 import com.hhplus.tdd.concert.domain.service.OutboxService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
